manage concurrent faults for a given page
The existing demand paging fault mechanism was not properly handling the case
when a fault for a page occurred while another fault for the same page was
being processed. The effects of such races were being masked by the fact that
the x86 page table code tacitly allows mappings to be overwritten. A mapping
set up as a result of a first fault could be displaced by a new mapping set up
by a subsequent fault. To rememdy this, a "pending_fault" entry is now created
for each virtual page address of a demand page fault. When a pending_fault
exists for a page that caused a fault, the faulting thread (or kernel context)
is added to the pending_fault as a dependent waiter. On completion of the page
table flush (tlb shootdown) associated with the mapping created, any
threads (and/or kernel context) that faulted on the page are resumed. To
enable such completions, the functions map_with_complete and
update_map_flags_with_complete have been added. Internally, flush completions
are now status_handlers rather than thunks.